Adjust mapcache a bit for our needs
We use mapcache for three descriptors in the request:
head, header (first desc) and status (last desc). Which sizes
known to be less than a page size. We do not use mapcache
for iovec descriptors, because this is sub-optimal due to
the iovec allocation nature at the guest size.

While here, refactor a bit.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>