Don't call xc_domain_set_target()
It is not needed for the backend running in Dom0 in the first place.
It is only needed for the backend running in other domain (DomD for example).
But the usage of this requires an extra changes to XSM policy (dummy)
which are unlikely correct. The XEN_DOMCTL_set_target implementation holds
reference on target domain which doesn't allow to completely destroy it.
So we need to check if the reference are hold by someone and drop it
when destroying target domain.

The correct way is to write our own XSM policy (FLASK) to allow
device model (running in other than dom0 domain) to perform operations
on its target domain rather than tweeking Xen itself.

So drop this call completely.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>