Unmap virtio descriptors right after stop using them
This wasn't an issue with using foreign mapping so far where
iovecs can remain mapped even after the request is completed
and nobody complains. With the grant mapping we need to unmap
iovecs right after stop using them and before signaling a frontent
that the responce is ready. Otherwise the frontend will complain
that corresponding grant refs are still in use when trying to
release them.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>