Fix build with MAP_IN_ADVANCE option
After Xen 4.18 some API was changed and the xc_domain_getinfo function
was removed. It was replaced by the xc_domain_getinfo_single function.
To use virtio-disk with the MAP_IN_ADVANCE option on newer Xen versions,
the xenctrl_get_dom_mem function needs to be updated to align with
the new API.

Signed-off-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>